requirements.txt
gradio>=3.50
streamlit>=1.39
torch
torchvision
huggingface_hub
Pillow
matplotlib
numpy
